Move menus functions to wp sub-folder
