+++ lib/wordpress/menus/getMenus.js
@@ -1,4 +1,4 @@
-import filterMenusByLocation from '@/functions/filterMenusByLocation'
+import filterMenusByLocation from '@/functions/wordpress/menus/filterMenusByLocation'
 
 // Define default menu locations.
 export const menuLocations = ['primary-menu', 'footer-menu', 'mobile-menu']
